backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
stripe == 4.2.0
python-dotenv